+++ R/QuadratiK-package.R
@@ -59,7 +59,8 @@
 #' The `QuadratiK` package is also available in Python on PyPI 
 #' <https://pypi.org/project/QuadratiK/> and also as a Dashboard application.
 #' Usage instruction for the Dashboard can be found at 
-#' <https://quadratik.readthedocs.io/en/latest/user_guide/dashboard_application_usage.html>.
+#' <https://quadratik.readthedocs.io/en/latest/user_guide/
+#' dashboard_application_usage.html>.
 #' 
 #' @author 
 #' Giovanni Saraceno, Marianthi Markatou, Raktim Mukhopadhyay, Mojgan Golzy