+++ R/pkbd_functions.R
@@ -3,7 +3,7 @@
 #' 
 #' @description 
 #' The Poisson kernel-based densities are based on the normalized Poisson kernel
-#' and are defined on the \eqn{d}-dimensional unit sphere. Given a vector 
+#' and are defined on the \eqn{(d-1)}-dimensional unit sphere. Given a vector 
 #' \eqn{\mathbf{\mu} \in \mathcal{S}^{d-1}}, where \eqn{\mathcal{S}^{d-1}= 
 #' \{x \in \mathbb{R}^d : ||x|| = 1\}}, and a parameter \eqn{\rho} such that 
 #' \eqn{0 < \rho < 1}, the probability density function of a \eqn{d}-variate 
@@ -31,8 +31,8 @@
 #' simulating from the PKBD, using angular central Gaussian envelops 
 #' (\code{rejacg}) or using the projected Saw distributions (\code{rejpsaw}).
 #'  
-#' @param x Matrix (or data.frame) of data point on the sphere 
-#'          \eqn{\mathcal{S}^{d-1}}, with \eqn{d \ge 2}.
+#' @param x \eqn{n \times d}-matrix (or data.frame) of \eqn{n} data point on the
+#'          sphere \eqn{\mathcal{S}^{d-1}}, with \eqn{d \ge 2}.
 #' @param mu Location parameter with same length as the rows of \code{x}.
 #' @param rho Concentration parameter, with \eqn{0 \le} \code{rho} \eqn{< 1}.
 #' @param logdens Logical; if 'TRUE', densities are returned in logarithmic
@@ -52,7 +52,7 @@
 #' pkbd_dat <- rpkb(10, c(0.5,0), 0.5)
 #' 
 #' # Calculate the PKBD density values
-#' dens_val <- dpkb(pkbd_dat$x, c(0.5,0.5),0.5)
+#' dens_val <- dpkb(pkbd_dat, c(0.5,0.5),0.5)
 #'
 #' @srrstats {G1.4} roxigen2 is used
 #' @srrstats {G2.0,G2.0a} check input mu 
@@ -126,9 +126,8 @@ dpkb <- function(x, mu, rho, logdens = FALSE) {
 #' 
 #' @details
 #' The number of observations generated is determined by \code{n} for 
-#' \code{rpkb()}. This function returns a list with the matrix of generated 
-#' observations \code{x}, the number of tries \code{numTries} and the number of
-#' acceptances \code{numAccepted}.
+#' \code{rpkb()}. This function returns the \eqn{(n \times d)}-matrix of 
+#' generated \eqn{n} observations on \eqn{\mathcal{S}^{(d-1)}}.
 #' 
 #' A limitation of the \code{rejvmf} is that the method does not ensure the
 #' computational feasibility of the sampler for \eqn{\rho} approaching 1.
@@ -227,7 +226,7 @@ rpkb <- function(n, mu, rho, method = 'rejacg',
       
    }
    
-   return(res)
+   return(res$x)
 }
 #'
 #' Random generation with vonMises distribution evelopes.