+++ tests/testthat/test-dpkb.R
@@ -24,36 +24,36 @@ test_that("Density of PKBD", {
    
    # Check the errors
    # mu
-   expect_error(dpkb(pkbd_dat$x, mu = c(1), rho = 0.8), 
+   expect_error(dpkb(pkbd_dat, mu = c(1), rho = 0.8), 
                 'mu must have length >= 2',fixed=TRUE)
-   expect_error(dpkb(pkbd_dat$x, mu = c(0,0,0), rho = 0.8), 
+   expect_error(dpkb(pkbd_dat, mu = c(0,0,0), rho = 0.8), 
                 'Input argument mu cannot be a vector of zeros',fixed=TRUE)
    
    expect_error(dpkb(rnorm(10), mu = c(0,0,0), rho = 0.8), 
                 "x must be a matrix or a data.frame",fixed=TRUE)
    expect_error(dpkb(matrix(rnorm(10),ncol=1), mu = c(0,0,0), rho = 0.8), 
                 'x must have dimension >= 2',fixed=TRUE)
-   expect_error(dpkb(pkbd_dat$x, mu = c(1,0), rho = 0.8), 
+   expect_error(dpkb(pkbd_dat, mu = c(1,0), rho = 0.8), 
                 'number of rows of x must be equal to the length of mu',
                 fixed=TRUE)
    
-   expect_error(dpkb(pkbd_dat$x, mu = c(1,0,0), rho = 2), 
+   expect_error(dpkb(pkbd_dat, mu = c(1,0,0), rho = 2), 
                 'Input argument rho must be within [0,1)',fixed=TRUE)
    
    #------------------------------------------------------
    ## Generating data point on the Sphere
    ## and computing the densities
-   den <- dpkb(pkbd_dat$x, mu, rho)
+   den <- dpkb(pkbd_dat, mu, rho)
    
    expect_equal(dim(den),c(size,1))
    expect_false(any(den<0))
    
-   den <- dpkb(data.frame(pkbd_dat$x), mu, rho)
+   den <- dpkb(data.frame(pkbd_dat), mu, rho)
    
    expect_equal(dim(den),c(size,1))
    expect_false(any(den<0))
    
-   log_den <- dpkb(pkbd_dat$x, mu, rho, logdens = TRUE)
+   log_den <- dpkb(pkbd_dat, mu, rho, logdens = TRUE)
    
    expect_equal(dim(log_den),c(size,1))
 })
@@ -83,7 +83,7 @@ test_that("tests - Density of PKBD", {
    size <- 100
    rho <- 0.8
    mu <- c(1,0,0)
-   pkbd_dat <- rpkb(size, mu = mu, rho = rho, method = "rejvmf")$x
+   pkbd_dat <- rpkb(size, mu = mu, rho = rho, method = "rejvmf")
    
    # Check the errors
    # x
@@ -100,27 +100,27 @@ test_that("tests - Density of PKBD", {
                 'x must have dimension >= 2',fixed=TRUE)
    
    pkbd_dat <- rpkb(size, mu = mu, rho = rho, method = "rejvmf")
-   expect_error(dpkb(pkbd_dat$x, mu = c(1,0), rho = 0.8), 
+   expect_error(dpkb(pkbd_dat, mu = c(1,0), rho = 0.8), 
                 'number of rows of x must be equal to the length of mu',
                 fixed=TRUE)
    
-   expect_error(dpkb(pkbd_dat$x, mu = c(1,0,0), rho = 2), 
+   expect_error(dpkb(pkbd_dat, mu = c(1,0,0), rho = 2), 
                 'Input argument rho must be within [0,1)',fixed=TRUE)
    
    #------------------------------------------------------
    ## Generating data point on the Sphere
    ## and computing the densities
-   den <- dpkb(pkbd_dat$x, mu, rho)
+   den <- dpkb(pkbd_dat, mu, rho)
    
    expect_equal(dim(den),c(size,1))
    expect_false(any(den<0))
    
-   den <- dpkb(data.frame(pkbd_dat$x), mu, rho)
+   den <- dpkb(data.frame(pkbd_dat), mu, rho)
    
    expect_equal(dim(den),c(size,1))
    expect_false(any(den<0))
    
-   log_den <- dpkb(pkbd_dat$x, mu, rho, logdens = TRUE)
+   log_den <- dpkb(pkbd_dat, mu, rho, logdens = TRUE)
    
    expect_equal(dim(log_den),c(size,1))
 })